refactor(pos) :  패키지 - context별로 추가
