+++ src/main/java/kitchenpos/order/common/domain/OrderLineItem.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain;
+package kitchenpos.order.common.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
