+++ src/main/java/kitchenpos/order/eatinorder/domain/JpaOrderTableRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain;
+package kitchenpos.order.eatinorder.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
