+++ src/main/java/kitchenpos/order/eatinorder/ui/OrderTableRestController.java
@@ -1,7 +1,7 @@
-package kitchenpos.order.ui;
+package kitchenpos.order.eatinorder.ui;
 
-import kitchenpos.order.application.OrderTableService;
-import kitchenpos.order.domain.OrderTable;
+import kitchenpos.order.eatinorder.application.OrderTableService;
+import kitchenpos.order.eatinorder.domain.OrderTable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
