+++ src/test/java/kitchenpos/Fixtures.java
@@ -3,7 +3,8 @@ package kitchenpos;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuProduct;
-import kitchenpos.order.domain.*;
+import kitchenpos.order.common.domain.*;
+import kitchenpos.order.eatinorder.domain.OrderTable;
 import kitchenpos.product.domain.Product;
 
 import java.math.BigDecimal;
